mlproject/requirements.txt
flask
flask-cors
torch
torchvision
pillow
matplotlib
numpy
